Add files via upload
The purpose of this project is to ease the 
management of a Library. It uses Python
programming with SQL connectivity concepts to 
make the work easier and support the smooth 
functioning of a Library.
This is Python Projects on Library Management 
system, which provided a lot of facility to their 
member to manage their activity easily.
The objective and scope of my Project Library
Management system is to record the details various 
activities of member. It will simplify the task and 
reduce the paper work.